pyproject.toml
[project]
name = "medic-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Quentin Delignon",email = "quentin.delignon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-ollama (>=0.3.5,<0.4.0)",
    "langchain-postgres (>=0.0.15,<0.0.16)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
